fix gaslimit long

+++ src/main/java/com/github/ontio/core/transaction/Transaction.java
@@ -174,8 +174,8 @@ public abstract class Transaction extends Inventory {
         json.put("Version", (int) version);
         json.put("Nonce", nonce& (Integer.MAX_VALUE*2-1));
         json.put("TxType", txType.value() & (Byte.MAX_VALUE*2-1));
-        json.put("GasPrice",gasPrice& (Integer.MAX_VALUE*2-1));
-        json.put("GasLimit",gasLimit& (Integer.MAX_VALUE*2-1));
+        json.put("GasPrice",gasPrice& (Long.MAX_VALUE*2-1));
+        json.put("GasLimit",gasLimit& (Long.MAX_VALUE*2-1));
         json.put("Payer",payer.toBase58());
         json.put("Attributes", Arrays.stream(attributes).map(p -> p.json()).toArray(Object[]::new));
         json.put("Sigs", Arrays.stream(sigs).map(p -> p.json()).toArray(Object[]::new));
